backend/requirements.txt
# NCERT Doubt-Solver FastAPI Backend
# Requirements for Python 3.12+

# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.0
pydantic-settings==2.1.0

# Async HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# MongoDB (Async)
motor==3.3.2
pymongo==4.6.2
dnspython==2.6.1

# Vector Database
pinecone>=5.0.0
sentence-transformers==2.3.1

# AI/LLM
openai==1.12.0
google-generativeai==0.3.2

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.2.0

# OCR
pytesseract==0.3.10

# Data Processing
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
tqdm==4.66.2

# Redis Cache (Optional)
redis==5.0.1

# Background Tasks
celery==5.3.4

# Production
gunicorn==21.2.0

# Development
pytest==8.0.0
pytest-asyncio==0.23.0
